+++ src/BugsnagUnity/SessionTracker.cs
@@ -9,7 +9,7 @@ namespace BugsnagUnity
     {
         void StartSession();
 
-        void StopSession();
+        void PauseSession();
 
         bool ResumeSession();
 
@@ -68,7 +68,7 @@ namespace BugsnagUnity
             }
         }
 
-        public void StopSession()
+        public void PauseSession()
         {
             var session = _currentSession;